# backend/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "BeltSectionRequest": {
        "additionalProperties": false,
        "properties": {
          "cord_depth_from_outer_m": {
            "minimum": 0.0,
            "title": "Cord Depth From Outer M",
            "type": "number"
          },
          "height_m": {
            "exclusiveMinimum": 0.0,
            "title": "Height M",
            "type": "number"
          },
          "inner_width_m": {
            "exclusiveMinimum": 0.0,
            "title": "Inner Width M",
            "type": "number"
          },
          "outer_width_m": {
            "exclusiveMinimum": 0.0,
            "title": "Outer Width M",
            "type": "number"
          }
        },
        "required": [
          "height_m",
          "outer_width_m",
          "inner_width_m",
          "cord_depth_from_outer_m"
        ],
        "title": "BeltSectionRequest",
        "type": "object"
      },
...
      "EndpointRadiiGeometryStudyRequest": {
        "additionalProperties": false,
        "properties": {
          "context": {
            "$ref": "#/components/schemas/GeometryDesignContextRequest"
          },
          "field_sampling": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GeometryFieldSamplingRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "minimum_primary_wrap_angle_rad": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minimum Primary Wrap Angle Rad"
          },
          "minimum_secondary_wrap_angle_rad": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minimum Secondary Wrap Angle Rad"
          },
          "primary_outer_radius_at_zero_shift_m": {
            "exclusiveMinimum": 0.0,
            "title": "Primary Outer Radius At Zero Shift M",
            "type": "number"
          },
          "sample_count": {
            "default": 301,
            "maximum": 4001.0,
            "minimum": 2.0,
            "title": "Sample Count",
            "type": "integer"
          },
          "secondary_outer_radius_at_zero_shift_m": {
            "exclusiveMinimum": 0.0,
            "title": "Secondary Outer Radius At Zero Shift M",
            "type": "number"
          }
        },
        "required": [
          "context",
          "primary_outer_radius_at_zero_shift_m",
          "secondary_outer_radius_at_zero_shift_m"
        ],
        "title": "EndpointRadiiGeometryStudyRequest",
        "type": "object"
      },
      "GeometryDesignContextRequest": {
        "additionalProperties": false,
        "properties": {
          "belt": {
            "$ref": "#/components/schemas/BeltSectionRequest"
          },
          "belt_outer_length_m": {
            "exclusiveMinimum": 0.0,
            "title": "Belt Outer Length M",
            "type": "number"
          },
          "deadzone_shift_m": {
            "minimum": 0.0,
            "title": "Deadzone Shift M",
            "type": "number"
          },
          "max_shift_m": {
            "minimum": 0.0,
            "title": "Max Shift M",
            "type": "number"
          },
          "sheave_half_angle_rad": {
            "exclusiveMaximum": 1.5707963267948966,
            "exclusiveMinimum": 0.0,
            "title": "Sheave Half Angle Rad",
            "type": "number"
          }
        },
        "required": [
          "belt",
          "belt_outer_length_m",
          "sheave_half_angle_rad",
          "deadzone_shift_m",
          "max_shift_m"
        ],
        "title": "GeometryDesignContextRequest",
        "type": "object"
      },
      "GeometryFieldSamplingRequest": {
        "additionalProperties": false,
        "properties": {
          "primary_outer_radius_m": {
            "items": {
              "type": "number"
            },
            "minItems": 2,
            "title": "Primary Outer Radius M",
            "type": "array"
          },
          "secondary_outer_radius_m": {
            "items": {
              "type": "number"
            },
            "minItems": 2,
            "title": "Secondary Outer Radius M",
            "type": "array"
          }
        },
        "required": [
          "primary_outer_radius_m",
          "secondary_outer_radius_m"
        ],
        "title": "GeometryFieldSamplingRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StaticStudyResponse": {
        "additionalProperties": false,
        "properties": {
          "study": {
            "additionalProperties": true,
            "title": "Study",
            "type": "object"
          }
        },
        "required": [
          "study"
        ],
        "title": "StaticStudyResponse",
        "type": "object"
      },
      "TargetRatiosGeometryStudyRequest": {
        "additionalProperties": false,
        "properties": {
          "context": {
            "$ref": "#/components/schemas/GeometryDesignContextRequest"
          },
          "field_sampling": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GeometryFieldSamplingRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "maximum_ratio": {
            "exclusiveMinimum": 0.0,
            "title": "Maximum Ratio",
            "type": "number"
          },
          "minimum_primary_wrap_angle_rad": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minimum Primary Wrap Angle Rad"
          },
          "minimum_ratio": {
            "exclusiveMinimum": 0.0,
            "title": "Minimum Ratio",
            "type": "number"
          },
          "minimum_secondary_wrap_angle_rad": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minimum Secondary Wrap Angle Rad"
          },
          "sample_count": {
            "default": 301,
            "maximum": 4001.0,
            "minimum": 2.0,
            "title": "Sample Count",
            "type": "integer"
          }
        },
        "required": [
          "context",
          "maximum_ratio",
          "minimum_ratio"
        ],
        "title": "TargetRatiosGeometryStudyRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "HTTP adapter around CINDER's public contracts. All CVT mechanics remain in CINDER; this service owns transport, runs, and presets.",
    "title": "CVT Simulator API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/studies/geometry/endpoint-radii": {
      "post": {
        "operationId": "geometry_endpoint_radii_api_v1_studies_geometry_endpoint_radii_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EndpointRadiiGeometryStudyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaticStudyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Geometry Endpoint Radii",
        "tags": [
          "studies"
        ]
      }
    },
    "/api/v1/studies/geometry/target-ratios": {
      "post": {
        "operationId": "geometry_target_ratios_api_v1_studies_geometry_target_ratios_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TargetRatiosGeometryStudyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaticStudyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Geometry Target Ratios",
        "tags": [
          "studies"
        ]
      }
    }
  }
}